+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.11
   - Add krb5-kdc-ldap with doc examples #171
+  - Fix NFS issue #169
 
 ## 1.1.10
 Fix: