+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.11
   - Add krb5-kdc-ldap with doc examples #171
   - Fix NFS issue #169
+  - Add support of readonly user in custom bootstrap ldif #162
 
 ## 1.1.10
 Fix: